+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add new `error` props on SelectWithState.
 - Fix: Change arrow-zone size on tiny select.
 
 ## [8.5.0] - 2017-03-14